catalystcount/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-allauth==64.2.0
djangorestframework==3.15.2
idna==3.10
jwt==1.3.1
pip==21.2.4
pycparser==2.22
requests==2.32.3
setuptools==58.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
